requirements.txt
keystoneauth1
python-swiftclient
